Move run file resolution out of interfaces
