+++ param_decomp/experiments/resid_mlp/models.py
@@ -13,9 +13,10 @@ from torch import Tensor, nn
 
 from param_decomp.base_config import BaseConfig
 from param_decomp.configs import ScheduleConfig, migrate_to_lr_schedule_config
-from param_decomp.interfaces import LoadableModule, resolve_run_files
+from param_decomp.interfaces import LoadableModule
 from param_decomp.param_decomp_types import ModelPath
 from param_decomp.utils.module_utils import init_param_
+from param_decomp.utils.run_files import resolve_run_files
 
 
 class ResidMLPModelConfig(BaseConfig):
