+++ param_decomp/experiments/tms/models.py
@@ -10,8 +10,9 @@ from torch.nn import functional as F
 
 from param_decomp.base_config import BaseConfig
 from param_decomp.configs import ScheduleConfig, migrate_to_lr_schedule_config
-from param_decomp.interfaces import LoadableModule, resolve_run_files
+from param_decomp.interfaces import LoadableModule
 from param_decomp.param_decomp_types import ModelPath
+from param_decomp.utils.run_files import resolve_run_files
 
 
 class TMSModelConfig(BaseConfig):
